layers: Split queue_state.{h,cpp} into multiple files
The classes in this file, especially vvl::Semaphore, have gotten
tricksy enough to justify having their own files.
