+++ BUILD.gn
@@ -187,6 +187,8 @@ vvl_sources = [
   "layers/state_tracker/device_memory_state.cpp",
   "layers/state_tracker/device_memory_state.h",
   "layers/state_tracker/device_state.h",
+  "layers/state_tracker/fence_state.cpp",
+  "layers/state_tracker/fence_state.h",
   "layers/state_tracker/image_layout_map.cpp",
   "layers/state_tracker/image_layout_map.h",
   "layers/state_tracker/image_state.cpp",
@@ -204,6 +206,8 @@ vvl_sources = [
   "layers/state_tracker/render_pass_state.cpp",
   "layers/state_tracker/render_pass_state.h",
   "layers/state_tracker/sampler_state.h",
+  "layers/state_tracker/semaphore_state.cpp",
+  "layers/state_tracker/semaphore_state.h",
   "layers/state_tracker/shader_instruction.cpp",
   "layers/state_tracker/shader_instruction.h",
   "layers/state_tracker/shader_module.cpp",
